redo entire project structure
The aim of the project is simply to serve as base for the "mentirinha"
app. As such, the name of the project is now a generic "sample_project"
instead of the main name, and the app is now called "mentirinha" instead
of "core". Good thing it's still at the very beginning.